% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsRDM.r
\docType{package}
\name{bbsRDM}
\alias{bbsRDM}
\alias{bbsRDM-package}
\title{bbsRDM: A package for calculating various 'regime detection' measures.}
\description{
The bbsRDM package provides source code for calculating multiple methods that are used in the ecological regime shift literatures.
foo, bar and baz.
}
\section{bbsRDM functions}{

}

