% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdsToFeathers.R
\name{birdsToFeathers}
\alias{birdsToFeathers}
\title{Save BBS dataframe as a feather file}
\usage{
birdsToFeathers(dataIn, newDir, filename)
}
\arguments{
\item{dataIn}{The BBS data to save.}

\item{newDir}{Where to save the BBS feathers.}

\item{filename}{Name of the new filename (e.g., 'arizona.zip'). This function will replace .zip with .feather}
}
\description{
Save BBS dataframe as a feather file
}
