% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distanceTravelled.R
\name{calculate_distanceTravelled}
\alias{calculate_distanceTravelled}
\title{Calculate the `distance travelled` by the entire system.}
\usage{
calculate_distanceTravelled(dataIn, derivs = T, print = T)
}
\arguments{
\item{dataIn}{input dataframe}

\item{derivs}{logical (default TRUE). calculate derivatives as well}

\item{print}{logical (default TRUE). print output}
}
\description{
Also calculates the velocity and acceleration of the entire system over the time series.
}
\description{
Calculate the `distance travelled` by the entire system. Also calculates the velocity and acceleration of the entire system over the time series.
}
