% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataBBS.R
\name{getDataBBS}
\alias{getDataBBS}
\title{Download USGS Breeding Bird Survey data}
\usage{
getDataBBS(file = file,
  dir = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/States/",
  year = NULL, aou = NULL, countrynum = NULL, states = NULL)
}
\arguments{
\item{file}{One file names ("stateX.zip"). Preferably download a single state at a time, otherwise run time will take >1 minutes.}

\item{dir}{URL to the StatesFiles.}

\item{year}{Vector of years. Default = NULL (all years).}

\item{aou}{Vector of AOU #s Default = NULL (all species).}

\item{countrynum}{Vector of country ID #'s. Default = NULL (all countryNums).}

\item{states}{Vector of state names Default = NULL (all states).}
}
\description{
This function was adapted from **oharar/rBBS** package.
}
