% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMilBases.R
\name{getMilBases}
\alias{getMilBases}
\title{Get military installation shapefile from online data repo and save to}
\usage{

  getMilBases(shploc = "http://www.acq.osd.mil/eie/Downloads/DISDI/installations_ranges.zip",
  shpfile = "FY18_MIRTA_Points")
}
\arguments{
\item{shploc}{URL location for.zip file}

\item{shpfile}{Name of the file to upload}
}
\value{
shp A shapefile with points designating U.S. military bases.
}
\description{
Get military installation shapefile from online data repo and save to
}
