% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSppListBBS.R
\name{getSppListBBS}
\alias{getSppListBBS}
\title{Import the species list used by the BBS from website or local disk.}
\usage{

  getSppListBBS(file = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/SpeciesList.txt",
  skip = 7)
}
\arguments{
\item{file}{A character string for location for the species list (as .txt).}

\item{skipEmpty}{Numeric, default = 7. The number of lines to skip. This may need to be updated if the file on webpage changes.}
}
\description{
Import the species list used by the BBS from website or local disk.
}
