% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBirdFeathers.R
\name{loadBirdFeathers}
\alias{loadBirdFeathers}
\title{Load the BBS data feathers into R.}
\usage{
loadBirdFeathers(newDir, filename)
}
\arguments{
\item{newDir}{Where the BBS feathers are saved.}

\item{filename}{Name of the feather filename (e.g., 'arizona.zip' or 'arizona'). This function will replace .zip with .feather when necessary.}
}
