% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFunMassBBS.R
\name{mergeFunMassBBS}
\alias{mergeFunMassBBS}
\title{Merge functional group and mass data for species with BBS counts}
\usage{
mergeFunMassBBS(bbsData, funMass, printMissing = T)
}
\arguments{
\item{bbsData}{The input bbsData.}

\item{funMass}{The funMass list (from `funcMass()`).}

\item{printMissing}{Logical. Prints to screen the missing species.}
}
\description{
Merge functional group and mass data for species with BBS counts
}
