% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linePlots.R
\name{sort.year.line}
\alias{sort.year.line}
\title{Plot a single transect over multiple years, with one metricTYpe.}
\usage{
\method{sort}{year.line}(df, metric.ind, year.ind, dirID.ind, direction,
  scale = TRUE, center = TRUE, min.data = 5)
}
\description{
Plot a single transect over multiple years, with one metricTYpe.
}
