% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByAOU.R
\name{subsetByAOU}
\alias{subsetByAOU}
\title{Subset the BBS data by species, functional traits, and/or body mass.}
\usage{
subsetByAOU(myData, subset.by = c("keep.diurnal", "remove.fowl",
  "remove.shorebirds", "remove.shoreWaderFowl"))
}
\arguments{
\item{myData}{A data frame including the column "aou".}

\item{subset.by}{One or more of   c("keep.diurnal", 'remove.fowl', "remove.shorebirds",  'remove.shoreWaderFowl').}

\item{mass}{Logical. Retrieves body mass information (Dunning reference).}
}
\description{
Subset the BBS data by species, functional traits, and/or body mass.
}
