% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetrics.R
\name{calculateMetrics}
\alias{calculateMetrics}
\title{Calculate regime detection metrics}
\usage{
calculateMetrics(dataIn, metrics.to.calc = c("distances", "ews"),
  min.samp.sites = 8, direction, yearInd, to.calc = c("EWS", "FI",
  "VI"))
}
\arguments{
\item{dataIn}{data frame with columns: sortVar (the sorting variable; latitude or longitude),  cellID (cell ID for the spatial grid),  variable (species),  value (count data).}

\item{metrics.to.calc}{One or more of c("distances", "ews")}

\item{min.samp.sites}{Minimum number of unique sites in the transect (or unique times along the time series) required to analyze the data. Most metrics can be calculated using three data points, although we do not nrecommend this.}

\item{direction}{Direction of the analysis (South-North or East-West)}
}
\description{
Calculates regime detection metrics across space or time. Calculates distance travelled, Fisher Information, Variance Index, Coefficient of Variation, mean, standard deviation, variance,skewness, and kurtosis. #' @param dataIn A data frame containing columns c(variable, time, value).
}
