% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importResults.R
\name{importResults}
\alias{importResults}
\title{Load the regime detection metric results (.feathers)}
\usage{
importResults(resultsDir, myPattern)
}
\arguments{
\item{resultsDir}{Where the results are stored.}

\item{myPattern}{Pattern for loading results files. One of "distance", "ews".}
}
\value{
A dataframe
}
\details{
Used after running calculate_distanceTravelled()? to make results available
  for visualization?
}
