% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByAOU.R
\name{subsetByAOU}
\alias{subsetByAOU}
\title{Subset the BBS data by species, functional traits, and/or body mass.}
\usage{
subsetByAOU(myData, subset.by = c("remove.fowl", "remove.shorebirds",
  "remove.shoreWaderFowl"))
}
\arguments{
\item{myData}{A data frame including the column "aou".}

\item{subset.by}{One or more of 'remove.fowl' (removes waterfowl), "remove.shorebirds" (removes shorebirds and waders),  'remove.shoreWaderFowl' (removes shorebirds, waders, and fowl).}

\item{mass}{Logical. Retrieves body mass information (Dunning reference).}
}
\description{
Subset the BBS data by species, functional traits, and/or body mass.
}
